+++ Classes/StrategiaGiocatore.cs
@@ -9,17 +9,17 @@ namespace Classes
     {
 
         public int Conteggio { get; set; }
-        public double TrueCount { get; set; }
-        public double GetTrueCount(int NumCarte)
+        public decimal TrueCount { get; set; }
+        public decimal GetTrueCount(int NumCarte)
         {
-            double ret = Math.Round((double)(Conteggio / ((NumCarte / 52) == 0 ? 1 : ((double)NumCarte / 52))),2);
+            decimal ret = Math.Round((decimal)(Conteggio / ((NumCarte / 52) == 0 ? 1 : ((decimal)NumCarte / 52))),2);
             return ret;
         }
 
 
-        public abstract Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, double conteggio);
-        public abstract bool Assicurazione(Giocatore giocatore, double conteggio);
-        public abstract double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, double Conteggio);
+        public abstract Giocatore.Puntata Strategy(Giocatore giocatore, Mazziere mazziere, decimal conteggio);
+        public abstract bool Assicurazione(Giocatore giocatore, decimal conteggio);
+        public abstract decimal Puntata(Giocatore giocatore, decimal puntataMinima, decimal puntataBase, decimal Conteggio);
         public virtual int Conta(Carta carta)
         {
             return 0;
